refactor: SearchInput 컴포넌트 리팩토링 fixed #43
* action 관련 로직 useSearchInput으로 분리
